Deduplicate Wave outcome cards
